[REFACTOR] 함께 의논하면서 엔티티 수정
