+++ src/main/java/com/fav/daengnyang/domain/account/service/dto/request/ColorUpdateRequest.java
@@ -1,0 +1,14 @@
+package com.fav.daengnyang.domain.account.service.dto.request;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+public class ColorUpdateRequest {
+    private String newColor;
+}
