--- src/main/java/com/fav/daengnyang/domain/bankbook/service/dto/request/BankbookRequest.java
@@ -1,15 +1,0 @@
-package com.fav.daengnyang.domain.bankbook.service.dto.request;
-
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-@Getter
-@Setter
-@NoArgsConstructor
-public class BankbookRequest {
-    private String bankbookTitle;
-    private String bankbookImage;
-    private String bankbookColor;
-    private String accountTypeUniqueNo;
-}
