--- src/main/java/com/fav/daengnyang/domain/bankbook/service/dto/request/ColorUpdateRequest.java
@@ -1,14 +1,0 @@
-package com.fav.daengnyang.domain.bankbook.service.dto.request;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@Builder
-@NoArgsConstructor
-@AllArgsConstructor
-public class ColorUpdateRequest {
-    private String newColor;
-}
