--- src/main/java/com/fav/daengnyang/domain/bankbook/service/dto/response/BankbookCreateResponse.java
@@ -1,15 +1,0 @@
-package com.fav.daengnyang.domain.bankbook.service.dto.response;
-
-import lombok.Builder;
-import lombok.Data;
-import lombok.Getter;
-
-@Data
-@Getter
-@Builder
-public class BankbookCreateResponse {
-    private String bankbookNumber;
-    private String bankbookTitle;
-    private String bankbookImage;
-    private String bankbookColor;
-}
