--- src/main/java/com/fav/daengnyang/domain/bankbook/service/dto/response/BankbookResponse.java
@@ -1,13 +1,0 @@
-package com.fav.daengnyang.domain.bankbook.service.dto.response;
-
-import lombok.Builder;
-import lombok.Getter;
-
-@Getter
-@Builder
-public class BankbookResponse {
-    private String bankbookTitle;
-    private String bankbookNumber;
-    private String bankbookImage;
-    private String bankbookColor;
-}
