--- src/main/java/com/fav/daengnyang/domain/bookdata/entity/BookdataDetail.java
@@ -1,27 +1,0 @@
-package com.fav.daengnyang.domain.bookdata.entity;
-
-import com.fav.daengnyang.domain.bankbook.entity.Bankbook;
-import jakarta.persistence.*;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-import java.time.LocalDateTime;
-
-@Entity
-@Getter
-@Setter
-@NoArgsConstructor
-public class BookdataDetail {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long detailId;
-
-    private Integer amount;
-    private LocalDateTime createdDate;
-
-    @ManyToOne
-    @JoinColumn(name = "bankbook_id")
-    private Bankbook bankbook;
-}
