--- src/main/java/com/fav/daengnyang/domain/bookdata/service/dto/response/BookdataSummaryResponse.java
@@ -1,15 +1,0 @@
-package com.fav.daengnyang.domain.bookdata.service.dto.response;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-
-@Data
-@Builder
-@AllArgsConstructor
-public class BookdataSummaryResponse {
-    private Integer totalAmount;
-    private Long transactionCount;
-    private String apiBalance;
-    private String apiHistory;
-}
