--- src/main/java/com/fav/daengnyang/domain/deposit/controller/DepositController.java
@@ -1,4 +1,0 @@
-package com.fav.daengnyang.domain.deposit.controller;
-
-public class DepositController {
-}
