--- src/main/java/com/fav/daengnyang/domain/deposit/entity/Deposit.java
@@ -1,4 +1,0 @@
-package com.fav.daengnyang.domain.deposit.entity;
-
-public class Deposit {
-}
