--- src/main/java/com/fav/daengnyang/domain/deposit/repository/DepositRepository.java
@@ -1,4 +1,0 @@
-package com.fav.daengnyang.domain.deposit.repository;
-
-public interface DepositRepository {
-}
