--- src/main/java/com/fav/daengnyang/domain/deposit/service/dto/DepositService.java
@@ -1,4 +1,0 @@
-package com.fav.daengnyang.domain.deposit.service.dto;
-
-public class DepositService {
-}
