--- src/main/java/com/fav/daengnyang/domain/deposit/service/dto/request/dummy.txt
@@ -1,1 +1,0 @@
-this is dummy file.
